Convert users service to service tracing provider.
This convers the users service to use the service tracing provider.
