+++ services/users/pkg/command/server.go
@@ -13,13 +13,13 @@ import (
 	"github.com/owncloud/ocis/v2/ocis-pkg/ldap"
 	"github.com/owncloud/ocis/v2/ocis-pkg/registry"
 	"github.com/owncloud/ocis/v2/ocis-pkg/sync"
+	"github.com/owncloud/ocis/v2/ocis-pkg/tracing"
 	"github.com/owncloud/ocis/v2/ocis-pkg/version"
 	"github.com/owncloud/ocis/v2/services/users/pkg/config"
 	"github.com/owncloud/ocis/v2/services/users/pkg/config/parser"
 	"github.com/owncloud/ocis/v2/services/users/pkg/logging"
 	"github.com/owncloud/ocis/v2/services/users/pkg/revaconfig"
 	"github.com/owncloud/ocis/v2/services/users/pkg/server/debug"
-	"github.com/owncloud/ocis/v2/services/users/pkg/tracing"
 	"github.com/urfave/cli/v2"
 )
 
@@ -34,7 +34,7 @@ func Server(cfg *config.Config) *cli.Command {
 		},
 		Action: func(c *cli.Context) error {
 			logger := logging.Configure(cfg.Service.Name, cfg.Log)
-			err := tracing.Configure(cfg, logger)
+			traceProvider, err := tracing.GetServiceTraceProvider(cfg.Tracing, cfg.Service.Name)
 			if err != nil {
 				return err
 			}
@@ -63,6 +63,7 @@ func Server(cfg *config.Config) *cli.Command {
 				runtime.RunWithOptions(rCfg, pidFile,
 					runtime.WithLogger(&logger.Logger),
 					runtime.WithRegistry(reg),
+					runtime.WithTraceProvider(traceProvider),
 				)
 
 				return nil
@@ -80,7 +81,6 @@ func Server(cfg *config.Config) *cli.Command {
 				debug.Context(ctx),
 				debug.Config(cfg),
 			)
-
 			if err != nil {
 				logger.Info().Err(err).Str("server", "debug").Msg("Failed to initialize server")
 				return err
